Implement separate geometry parsing
Create a new geomtetry struct to store all the geometry data to avoid
having to use the X macros to query the values. Additionally, add the
color array to the geometry structure in order to have all the data
needed to display a notification in a single struct.

This commit is part of the draw refactor.